+++ src/models/outlets/outlets-request-context-response.model.ts
@@ -2,6 +2,16 @@ export interface OutletsRequestContextResponse {
   target?: string;
   isRootNodeHttps?: boolean;
   isConfigurationMode: boolean;
-  plugin?: any;
+  plugin?: PluginForOutlet;
   isPreviewActive: boolean;
+}
+
+interface PluginForOutlet {
+  name: string;
+  url: string;
+  optimalHeight?: string;
+  useShellSDK?: boolean;
+  isActive: boolean;
+  sandboxPolicies?: string[];
+  assignmentId?: string;
 }