+++ src/validation/schemas/feature-flag/get-feature-flag-request.v1.schema.ts
@@ -1,12 +1,36 @@
 export const getFeatureFlagRequest_v1_schema = {
-  type: 'object',
-  properties: {
-    key: {
-      type: 'string',
-    },
-    defaultValue: {
-      type: 'boolean',
-    },
+  $defs: {
+    payload: {
+      type: 'object',
+      properties: {
+        key: {
+          type: 'string'
+        },
+        defaultValue: {
+          type: 'boolean'
+        }
+      },
+      required: ['key', 'defaultValue']
+    }
   },
-  required: ['key', 'defaultValue'],
+  anyOf: [
+    {
+      type: 'object',
+      properties: {
+        key: {
+          type: 'string'
+        },
+        defaultValue: {
+          type: 'boolean'
+        }
+      },
+      required: ['key', 'defaultValue']
+    },
+    {
+      type: 'array',
+      items: {
+        $ref: '#/$defs/payload'
+      }
+    }
+  ]
 };