+++ src/validation/schemas/outlets/outlets-request-dynamic-context-response.v1.schema.ts
@@ -1,22 +1,59 @@
 export const outletsRequestDynamicContextResponse_v1_schema = {
+  $defs: {
+    plugin: {
+      type: 'object',
+      properties: {
+        name: {
+          type: 'string'
+        },
+        url: {
+          type: 'string'
+        },
+        optimalHeight: {
+          type: 'string'
+        },
+        useShellSDK: {
+          type: 'boolean'
+        },
+        isActive: {
+          type: 'boolean'
+        },
+        sandboxPolicies: {
+          type: 'array',
+          items: {
+            type: 'string'
+          }
+        },
+        assignmentId: {
+          type: 'string'
+        }
+      },
+      required: ['name', 'url', 'isActive', 'assignmentId']
+    }
+  },
   type: 'object',
   properties: {
     target: {
-      type: 'string',
+      type: 'string'
     },
     isRootNodeHttps: {
-      type: 'boolean',
+      type: 'boolean'
     },
     isConfigurationMode: {
-      type: 'boolean',
+      type: 'boolean'
     },
     areDynamicOutletsEnabled: {
-      type: 'boolean',
+      type: 'boolean'
     },
     isPreviewActive: {
-      type: 'boolean',
+      type: 'boolean'
     },
-    plugins: [],
+    plugins: {
+      type: 'array',
+      items: {
+        $ref: '#/$defs/plugin'
+      }
+    }
   },
-  required: [],
+  required: []
 };